Bug fixes, code improvments, faster.
Made two new functions:

1) printError --> Takes the string of the error as input and prints the
error in red colour.
2) userInput --> A function to take user input printing the according
message, simply because the proccess repeats.

--------------------------------------------------------------------
Two bugs where fixed both related to time.
Converting from the input string to the date and time is simplified and
doesnt get cut and converted from the string in every if statement. The
code should be cleaner.